refactor: rename jsi interfaces
