+++ packages/react-native-audio-api/src/AudioAPIModule/ModuleInterfaces.ts
@@ -1,4 +1,4 @@
-import type { ShareableWorkletCallback } from '../interfaces.jsi';
+import type { ShareableWorkletCallback } from '../jsi-interfaces';
 
 // Hint: Copied from react-native-worklets
 // Doesn't really matter what is inside, just need a unique type
