+++ packages/react-native-audio-api/src/core/AudioBufferBaseSourceNode.ts
@@ -2,7 +2,7 @@ import AudioParam from './AudioParam';
 import type BaseAudioContext from './BaseAudioContext';
 import { AudioEventSubscription } from '../events';
 import { EventTypeWithValue } from '../events/types';
-import { IAudioBufferBaseSourceNode } from '../interfaces.jsi';
+import { IAudioBufferBaseSourceNode } from '../jsi-interfaces';
 import AudioScheduledSourceNode from './AudioScheduledSourceNode';
 
 export default class AudioBufferBaseSourceNode extends AudioScheduledSourceNode {
