+++ packages/react-native-audio-api/src/core/AudioBufferSourceNode.ts
@@ -1,4 +1,4 @@
-import { IAudioBufferSourceNode } from '../interfaces.jsi';
+import { IAudioBufferSourceNode } from '../jsi-interfaces';
 import AudioBufferBaseSourceNode from './AudioBufferBaseSourceNode';
 import AudioBuffer from './AudioBuffer';
 import { InvalidStateError, RangeError } from '../errors';
