+++ packages/react-native-audio-api/src/core/AudioNode.ts
@@ -1,4 +1,4 @@
-import { IAudioNode } from '../interfaces.jsi';
+import { IAudioNode } from '../jsi-interfaces';
 import AudioParam from './AudioParam';
 import { ChannelCountMode, ChannelInterpretation } from '../types';
 import type BaseAudioContext from './BaseAudioContext';
