+++ packages/react-native-audio-api/src/core/BaseAudioContext.ts
@@ -3,7 +3,7 @@ import {
   InvalidStateError,
   NotSupportedError,
 } from '../errors';
-import { IBaseAudioContext } from '../interfaces.jsi';
+import { IBaseAudioContext } from '../jsi-interfaces';
 import { AudioWorkletRuntime, ContextState, DecodeDataInput } from '../types';
 import { assertWorkletsEnabled } from '../utils';
 import AnalyserNode from './AnalyserNode';
