+++ packages/react-native-audio-api/src/core/MediaElementAudioSourceNode.ts
@@ -1,6 +1,6 @@
 import AudioNode from './AudioNode';
 import type AudioContext from './AudioContext';
-import { IAudioContext } from '../interfaces.jsi';
+import { IAudioContext } from '../jsi-interfaces';
 import type {
   AudioTagHandle,
   InternalAudioTagHandle,
