+++ packages/react-native-audio-api/src/core/RecorderAdapterNode.ts
@@ -1,4 +1,4 @@
-import { IRecorderAdapterNode } from '../interfaces.jsi';
+import { IRecorderAdapterNode } from '../jsi-interfaces';
 import AudioNode from './AudioNode';
 import type BaseAudioContext from './BaseAudioContext';
