+++ packages/react-native-audio-api/src/core/WaveShaperNode.ts
@@ -1,7 +1,7 @@
 import AudioNode from './AudioNode';
 import type BaseAudioContext from './BaseAudioContext';
 import { InvalidStateError } from '../errors';
-import { IWaveShaperNode } from '../interfaces.jsi';
+import { IWaveShaperNode } from '../jsi-interfaces';
 import { WaveShaperOptions } from '../types';
 
 export default class WaveShaperNode extends AudioNode {
