+++ packages/react-native-audio-api/src/development/react/Audio/AudioFileSourceNode.ts
@@ -3,7 +3,7 @@ import type { EventEmptyType } from '../../../events/types';
 import type {
   IAudioFileSourceNode,
   IAudioScheduledSourceNode,
-} from '../../../interfaces.jsi';
+} from '../../../jsi-interfaces';
 import AudioScheduledSourceNode from '../../../core/AudioScheduledSourceNode';
 
 type AttachFileSourceOptions = {
